# static/swagger.json
{
    "openapi": "3.0.2",
    "info": {
      "title": "Investment Properties API",
      "description": "API for accessing investment property data with comprehensive filtering, sorting, and search capabilities.",
      "version": "1.0.0",
      "contact": {
        "email": "support@example.com"
      }
    },
    "servers": [
      {
        "url": "/",
        "description": "Current server"
      }
    ],
...
    "paths": {
...
      "/api/properties": {
        "get": {
          "tags": ["properties"],
          "summary": "Search for properties",
          "description": "Search for investment properties with various filters and sorting options",
          "operationId": "searchProperties",
          "parameters": [
            {
              "name": "zip_code",
              "in": "query",
              "description": "Filter by zip code",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "city",
              "in": "query",
              "description": "Filter by city name",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "state",
              "in": "query",
              "description": "Filter by state code",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "min_price",
              "in": "query",
              "description": "Minimum property price",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "max_price",
              "in": "query",
              "description": "Maximum property price",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "sort_by",
              "in": "query",
              "description": "Field to sort by",
              "schema": {
                "type": "string",
                "enum": ["investment_ranking", "price_per_sqft", "cap_rate", "cash_on_cash", "irr", "total_return", "list_price"],
                "default": "investment_ranking"
              }
            },
            {
              "name": "page",
              "in": "query",
              "description": "Page number for pagination",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1,
                "minimum": 1
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Results per page",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 20,
                "minimum": 1,
                "maximum": 100
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with property results",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PropertySearchResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - invalid parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/properties/{property_id}": {
        "get": {
          "tags": ["properties"],
          "summary": "Get property details",
          "description": "Get detailed information for a specific property",
          "operationId": "getPropertyDetail",
          "parameters": [
            {
              "name": "property_id",
              "in": "path",
              "description": "Unique property identifier",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with property details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PropertyDetailResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Property not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/properties/{property_id}/audit": {
        "get": {
          "tags": ["properties"],
          "summary": "Get property calculation audit",
          "description": "Get detailed calculation audit data for a specific property",
          "operationId": "getPropertyAudit",
          "parameters": [
            {
              "name": "property_id",
              "in": "path",
              "description": "Unique property identifier",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with property calculation audit data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PropertyAuditResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Property or audit data not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/cities": {
        "get": {
          "tags": ["locations"],
          "summary": "Get cities list",
          "description": "Get a list of available cities, optionally filtered by state",
          "operationId": "getCities",
          "parameters": [
            {
              "name": "state",
              "in": "query",
              "description": "Filter cities by state code",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with cities list",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CitiesResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/zipcodes": {
        "get": {
          "tags": ["locations"],
          "summary": "Get zip codes list",
          "description": "Get a list of available zip codes, optionally filtered by city and/or state",
          "operationId": "getZipcodes",
          "parameters": [
            {
              "name": "city",
              "in": "query",
              "description": "Filter zip codes by city name",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "state",
              "in": "query",
              "description": "Filter zip codes by state code",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with zip codes list",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ZipcodesResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/states": {
        "get": {
          "tags": ["locations"],
          "summary": "Get states list",
          "description": "Get a list of available states with property counts",
          "operationId": "getStates",
          "responses": {
            "200": {
              "description": "Successful response with states list",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/StatesResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/stats/market": {
        "get": {
          "tags": ["statistics"],
          "summary": "Get market statistics",
          "description": "Get market statistics by location type",
          "operationId": "getMarketStats",
          "parameters": [
            {
              "name": "type",
              "in": "query",
              "description": "Location type (city, zipcode, or state)",
              "schema": {
                "type": "string",
                "enum": ["city", "zipcode", "state"],
                "default": "city"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with market statistics",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MarketStatsResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - invalid location type",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "State": {
          "type": "object",
          "properties": {
            "state": {
              "type": "string",
              "example": "CA"
            },
            "property_count": {
              "type": "integer",
              "format": "int32",
              "example": 450
            }
          }
        },
...
        "PropertySearchResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "object",
              "properties": {
                "properties": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertySummary"
                  }
                },
                "pagination": {
                  "$ref": "#/components/schemas/PaginationInfo"
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "zip_code": {
                      "type": ["string", "null"],
                      "example": null
                    },
                    "city": {
                      "type": ["string", "null"],
                      "example": "New York"
                    },
                    "state": {
                      "type": ["string", "null"],
                      "example": "NY"
                    },
                    "min_price": {
                      "type": ["string", "null"],
                      "example": "500000"
                    },
                    "max_price": {
                      "type": ["string", "null"],
                      "example": "1000000"
                    },
                    "sort_by": {
                      "type": "string",
                      "example": "cap_rate"
                    }
                  }
                }
              }
            }
          }
        },
        "PropertyDetailResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "$ref": "#/components/schemas/PropertyDetail"
            }
          }
        },
        "PropertyAuditResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "$ref": "#/components/schemas/PropertyAudit"
            }
          }
        },
        "CitiesResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "object",
              "properties": {
                "cities": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/City"
                  }
                },
                "count": {
                  "type": "integer",
                  "format": "int32",
                  "example": 45
                }
              }
            }
          }
        },
        "ZipcodesResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "object",
              "properties": {
                "zipcodes": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Zipcode"
                  }
                },
                "count": {
                  "type": "integer",
                  "format": "int32",
                  "example": 15
                }
              }
            }
          }
        },
        "StatesResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "object",
              "properties": {
                "states": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/State"
                  }
                },
                "count": {
                  "type": "integer",
                  "format": "int32",
                  "example": 35
                }
              }
            }
          }
        },
...
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": false
            },
            "error": {
              "type": "string",
              "example": "Not found"
            },
            "message": {
              "type": "string",
              "example": "The requested resource does not exist"
            }
          }
        }
      }
    }
  }